+++ src/main/kotlin/blackjack/ui/InputView.kt
@@ -1,9 +1,5 @@
 package blackjack.ui
 
-typealias UserName = String
-typealias UserNames = List<UserName>
-typealias UserMore = Boolean
-
 class InputView(val inputProvider: () -> String? = { readln() }) {
     fun inputUserNames(): UserNames {
         println("게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)")
