+++ src/main/kotlin/blackjack/ui/ResultView.kt
@@ -1,13 +1,5 @@
 package blackjack.ui
 
-typealias CardRank = String
-typealias CardSuit = String
-typealias UserCards = Map<CardRank, List<CardSuit>>
-typealias Score = Int
-typealias RoundResult = Map<UserName, UserCards>
-typealias ViewResult = Map<UserName, Map<UserCards, Score>>
-typealias ViewSuits = Map<String, String>
-
 class ResultView(val outputProvider: (String) -> Unit = { println(it) }) {
     fun printUserCardCount(
         names: UserNames,
@@ -49,22 +41,6 @@ class ResultView(val outputProvider: (String) -> Unit = { println(it) }) {
     ) {
         cards.forEach { (cards, score) ->
             printResult(userName, cards, score)
-        }
-    }
-
-    companion object {
-        private val VIEW_SUIT: ViewSuits =
-            mapOf(
-                "HEART" to "하트",
-                "DIAMOND" to "다이아몬드",
-                "CLUB" to "클로버",
-                "SPADE" to "스페이드",
-            )
-
-        fun UserCards.toPrettyString(): String {
-            return this.entries.joinToString(separator = ", ") { (rank, suits) ->
-                suits.joinToString(separator = ", ") { "$rank${VIEW_SUIT[it]}" }
-            }
         }
     }
 }
